+++ chain/client/src/sync/state/downloader.rs
@@ -14,7 +14,6 @@ use near_primitives::state_sync::{ShardStateSyncResponseHeader, StatePartKey};
 use near_primitives::types::ShardId;
 use near_store::{DBCol, Store};
 use std::sync::Arc;
-use std::sync::atomic::{AtomicUsize, Ordering};
 use tokio_util::sync::CancellationToken;
 use tracing::Instrument;
 
@@ -29,8 +28,6 @@ pub(super) struct StateSyncDownloader {
     pub clock: Clock,
     pub store: Store,
     pub preferred_source: Arc<dyn StateSyncDownloadSource>,
-    pub fallback_source: Option<Arc<dyn StateSyncDownloadSource>>,
-    pub num_attempts_before_fallback: usize,
     pub header_validation_sender:
         AsyncSender<SpanWrapped<StateHeaderValidationRequest>, Result<(), near_chain::Error>>,
     pub runtime: Arc<dyn RuntimeAdapter>,
@@ -53,8 +50,6 @@ impl StateSyncDownloader {
         let store = self.store.clone();
         let validation_sender = self.header_validation_sender.clone();
         let preferred_source = self.preferred_source.clone();
-        let fallback_source = self.fallback_source.clone();
-        let num_attempts_before_fallback = self.num_attempts_before_fallback;
         let task_tracker = self.task_tracker.clone();
         let clock = self.clock.clone();
         let retry_backoff = self.retry_backoff;
@@ -67,22 +62,9 @@ impl StateSyncDownloader {
                 return Ok(header);
             }
 
-            let i = AtomicUsize::new(0); // for easier Rust async capture
             let attempt = || {
                 async {
-                    // We cannot assume that either source is infallible. We interleave attempts
-                    // to the available sources until one of them gives us the state successfully.
-                    let source = if fallback_source.is_some()
-                        && i.load(Ordering::Relaxed) >= num_attempts_before_fallback
-                    {
-                        i.store(0, Ordering::Relaxed);
-                        fallback_source.as_ref().unwrap().as_ref()
-                    } else {
-                        i.fetch_add(1, Ordering::Relaxed);
-                        preferred_source.as_ref()
-                    };
-
-                    let header = source
+                    let header = preferred_source
                         .download_shard_header(shard_id, sync_hash, handle.clone(), cancel.clone())
                         .await?;
                     // We cannot validate the header with just a Store. We need the Chain, so we queue it up
@@ -160,14 +142,11 @@ impl StateSyncDownloader {
         state_root: CryptoHash,
         num_state_parts: u64,
         part_id: u64,
-        num_prior_attempts: usize,
         cancel: CancellationToken,
     ) -> BoxFuture<'static, Result<(), near_chain::Error>> {
         let store = self.store.clone();
         let runtime_adapter = self.runtime.clone();
         let preferred_source = self.preferred_source.clone();
-        let fallback_source = self.fallback_source.clone();
-        let num_attempts_before_fallback = self.num_attempts_before_fallback;
         let clock = self.clock.clone();
         let task_tracker = self.task_tracker.clone();
         let retry_backoff = self.retry_backoff;
@@ -183,19 +162,7 @@ impl StateSyncDownloader {
             }
 
             let attempt = || async {
-                // We cannot assume that either source is infallible. We cycle attempts
-                // to the available sources until one of them gives us the state successfully.
-                let cycle_length = num_attempts_before_fallback + 1;
-                let use_fallback = fallback_source.is_some()
-                    && num_prior_attempts % cycle_length == num_attempts_before_fallback;
-
-                let source = if use_fallback {
-                    fallback_source.as_ref().unwrap().as_ref()
-                } else {
-                    preferred_source.as_ref()
-                };
-
-                let part = source
+                let part = preferred_source
                     .download_shard_part(
                         shard_id,
                         sync_hash,