+++ chain/indexer/src/lib.rs
@@ -202,6 +202,5 @@ pub fn indexer_init_configs(
         params.download_config_url.as_deref(),
         params.boot_nodes.as_deref(),
         params.max_gas_burnt_view,
-        None,
     )
 }